+++ src/pydiverse/pipedag/backend/table/sql/__init__.py
@@ -1,3 +1,8 @@
 from __future__ import annotations
 
-from .sql import SQLTableStore
+from .sql import SQLTableStore, TableReference
+
+__all__ = [
+    "SQLTableStore",
+    "TableReference",
+]